# Install net-tools on eoan