route command from this package.

Before fix 
```
11/13/2019 03:52:53 : [DEBUG] .\Tools\plink.exe -ssh -t -pw ****** -P 1111 root@52.250.107.240 ". ./utils.sh &> /dev/null && get_active_nic_name"

11/13/2019 03:52:55 : [DEBUG] . ./utils.sh &> /dev/null && get_active_nic_name executed successfully in 1.29 seconds.

11/13/2019 03:52:55 : [INFO ] CLIENT Synthetic NIC: ./utils.sh: line 2802: route: command not found

11/13/2019 03:52:55 : [INFO ] SERVER Synthetic NIC: ./utils.sh: line 2802: route: command not found
```

After fix
```
[LISAv2 Test Results Summary]
Test Run On           : 11/08/2019 15:09:15
ARM Image Under Test  : Canonical : UbuntuServer : 19.10-DAILY : latest
Initial Kernel Version: 5.3.0-1004-azure
Final Kernel Version  : 5.3.0-1006-azure
Total Test Cases      : 1 (1 Passed, 0 Failed, 0 Aborted, 0 Skipped)
Total Time (dd:hh:mm) : 0:1:42

   ID TestArea             TestCaseName                                                                TestResult TestDuration(in minutes) 
-------------------------------------------------------------------------------------------------------------------------------------------
    1 NETWORK              PERF-NETWORK-TCP-THROUGHPUT-MULTICONNECTION-NTTTCP-Synthetic                      PASS                94.66 
	Connections=1 : throughput=3.47Gbps cyclePerBytet=1.14 Avg_TCP_lat=261.357 
	Connections=2 : throughput=5.90Gbps cyclePerBytet=.95 Avg_TCP_lat=270.387 
	Connections=4 : throughput=10.22Gbps cyclePerBytet=.56 Avg_TCP_lat=6180.995 
	Connections=8 : throughput=16.77Gbps cyclePerBytet=.67 Avg_TCP_lat=343.335 
	Connections=16 : throughput=18.46Gbps cyclePerBytet=.70 Avg_TCP_lat=8924.620 
	Connections=32 : throughput=19.57Gbps cyclePerBytet=.87 Avg_TCP_lat=10847.185 
	Connections=64 : throughput=18.63Gbps cyclePerBytet=.94 Avg_TCP_lat=6459.296 
	Connections=128 : throughput=18.21Gbps cyclePerBytet=1.12 Avg_TCP_lat=6921.583 
	Connections=256 : throughput=17.70Gbps cyclePerBytet=1.67 Avg_TCP_lat=9749.508 
	Connections=512 : throughput=17.36Gbps cyclePerBytet=1.31 Avg_TCP_lat=14695.820 
	Connections=1024 : throughput=11.70Gbps cyclePerBytet=4.20 Avg_TCP_lat=10647.759 
	Connections=2048 : throughput=6.82Gbps cyclePerBytet=6.08 Avg_TCP_lat=13137.631 
	Connections=4096 : throughput=6.32Gbps cyclePerBytet=8.30 Avg_TCP_lat=29532.688 
	Connections=6144 : throughput=5.73Gbps cyclePerBytet=11.02 Avg_TCP_lat=34992.878 
	Connections=8192 : throughput=4.91Gbps cyclePerBytet=13.66 Avg_TCP_lat=36973.602 
	Connections=10240 : throughput=5.88Gbps cyclePerBytet=10.84 Avg_TCP_lat=49561.550 
```